pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clawductor"
version = "0.1.0"
description = "TUI orchestration runtime for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.60.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
clawductor = "clawductor.main:main"

[tool.hatch.build.targets.wheel]
packages = ["clawductor"]
